refactor(validation): extract shared utilities from validation.ts (#3594)
Extract 349 lines of shared utilities from the 4521-line validation.ts
into validation/helpers.ts:
- URL normalization (normalizeBaseUrl, resolveBaseUrl, etc.)
- Header builders (buildBearerHeaders, buildKeyHeaders, etc.)
- HTTP helpers (validationRead, validationWrite, directHttpsRequest)
- Error handling (toValidationErrorResult, isSecurityBlockError)
- Format constants (OPENAI_LIKE_FORMATS, GEMINI_LIKE_FORMATS)

validation.ts → validation/index.ts (4521 → 4236 lines).
Pure extraction, no behavioral changes. Part of #3594.
